Move kntest and rundk to the tools folder
