+++ tools/kntest/pkg/kubetest2/gke/commands.go
