+++ tools/kntest/pkg/metadata/command.go
