+++ tools/rundk/interactive/command.go
