+++ tools/rundk/interactive/command_test.go
